requirements.txt
tensorflow==2.13.0
transformers==4.35.0
numpy==1.24.3
pillow==10.0.0
scikit-learn==1.3.0
pyyaml==6.0.1
pandas==2.0.3
matplotlib==3.7.2
tqdm==4.65.0
